backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.2
openai==1.3.5
pydub==0.25.1
gitpython==3.1.40
websockets==11.0.3
aiohttp==3.8.6
transformers==4.35.2
torch==2.2.0
numpy==1.26.4
soundfile==0.12.1
librosa==0.10.1
pyaudio==0.2.11
coral-python==1.0.2
blackbox-ai==1.2.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
asyncio==3.4.3
websocket-client==1.6.4